- check for previous packages before remove them  - change the outputs

+++ src/mkxpi.sh
@@ -1,38 +1,42 @@
 #!/bin/bash
 
 function cleanBuild {
-  if [ -d $BUILD_DIR ]; then
+  if [ -d $BUILD_DIR ];then
     rm -rf $BUILD_DIR;
   fi
 }
 function buildXPI {
-  echo Creating SurfKeys installation package;
-  echo Extension jar package chrome.jar;
+  echo "Creating SurfKeys installation package";
   cd chrome;
-  echo "delete jar";
-  rm chrome.jar;
-  echo Build package surfkeys.jar
+  if [ -f chrome.jar ];then
+    rm chrome.jar;
+  fi;
+  echo "Build package surfkeys.jar";
   zip -r chrome.jar content/* -x \*CVS/\*;
   zip -r chrome.jar locale/* -x \*CVS/\*;
   zip -r chrome.jar skin/* -x \*CVS/\*;
 
   cd ..;
-  echo Build package surfkeys.xpi;
+  echo "Build package surfkeys.xpi";
   rm surfkeys.xpi;
   zip surfkeys.xpi chrome.manifest install.rdf chrome/chrome.jar defaults/preferences/prefs.js -x \*CVS/\*;
 
-  rm $START_DIR/../downloads/surfkeys_$SFFVER.xpi;
+  echo "Replace old XPIs with the new one";
+  if [ -f $START_DIR/../downloads/surfkeys_$SFFVER.xpi ];then
+    rm $START_DIR/../downloads/surfkeys_$SFFVER.xpi;
+  fi;
   cp surfkeys.xpi $START_DIR/../downloads/surfkeys_$SFFVER.xpi;
-  rm $START_DIR/surfkeys.xpi;
+  if [ -f $START_DIR/surfkeys.xpi ];then
+    rm $START_DIR/surfkeys.xpi;
+  fi;
   cp surfkeys.xpi $START_DIR/;
-
-  echo Done!
+  echo "Build finished!";
 }
 function setVersion {
   if [ `pwd` != $BUILD_DIR ];then
     cd $BUILD_DIR;
   fi
-  echo "set version to $SFVER";
+  echo "Set version to $SFVER";
   sed "s/###VERSION###/$SFVER/g" install.rdf > install.rdf.tmp;
   mv install.rdf.tmp install.rdf;
 }
@@ -49,9 +53,7 @@ TMP_DIR=/tmp;
 BUILD_DIR=$TMP_DIR/surfkeys_$SFVER;
 
 if [ -d $TMP_DIR ];then
-  if [ -e $BUILD_DIR ];then
-    cleanBuild;
-  fi
+  cleanBuild;
   echo "Copy current files to temp dir to build package";
   cp -R $START_DIR $BUILD_DIR;
   cd $BUILD_DIR;
